# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenDesk API
  description: >
    Sovereign, open-source office suite API. Provides document management,
    real-time collaboration, sharing, audit trails, workflow automation,
    cross-sovereign federation, local AI assistance, and verifiable data erasure.
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:5500
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT or dev-mode token passed as Bearer header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            type: object
      required: [error, issues]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]

    # --- Documents ---
...
    Reference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: integer
          nullable: true
        source:
          type: string
          nullable: true
        type:
          type: string
          enum: [article, book, chapter, website, other]
        doi:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        publisher:
          type: string
          nullable: true
    CreateReferenceBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        authors:
          type: array
          items:
            type: string
          default: []
        year:
          type: integer
          nullable: true
        source:
          type: string
          maxLength: 500
          nullable: true
        volume:
          type: string
          maxLength: 100
          nullable: true
        issue:
          type: string
          maxLength: 100
          nullable: true
        pages:
          type: string
          maxLength: 100
          nullable: true
        doi:
          type: string
          maxLength: 200
          nullable: true
        isbn:
          type: string
          maxLength: 50
          nullable: true
        url:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
        publisher:
          type: string
          maxLength: 500
          nullable: true
        type:
          type: string
          enum: [article, book, chapter, website, other]
          default: other
    UpdateReferenceBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        authors:
          type: array
          items:
            type: string
        year:
          type: integer
          nullable: true
        source:
          type: string
          maxLength: 500
          nullable: true
        volume:
          type: string
          maxLength: 100
          nullable: true
        issue:
          type: string
          maxLength: 100
          nullable: true
        pages:
          type: string
          maxLength: 100
          nullable: true
        doi:
          type: string
          maxLength: 200
          nullable: true
        isbn:
          type: string
          maxLength: 50
          nullable: true
        url:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
        publisher:
          type: string
          maxLength: 500
          nullable: true
        type:
          type: string
          enum: [article, book, chapter, website, other]

    # --- Sharing ---
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
        action:
          type: string
        actor_id:
          type: string
        hmac:
          type: string
        created_at:
          type: string
          format: date-time
    AuditProof:
      type: object
      properties:
        documentId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditEntry"
        hmacSecret:
          type: string

    # --- Workflow ---
...
security:
  - BearerAuth: []
paths:
...
  /api/references:
    get:
      tags: [References]
      summary: List references
      description: List references in the library. Optionally filter by documentId.
      parameters:
        - name: documentId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of references
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reference"
        "403":
          description: No read access to reference library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [References]
      summary: Create a reference
      description: Add a new reference to the library.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferenceBody"
      responses:
        "201":
          description: Created reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "403":
          description: No write access to reference library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/references/lookup/doi/{doi}:
    get:
      tags: [References]
      summary: Lookup reference by DOI
      description: Fetch reference metadata from CrossRef by DOI.
      parameters:
        - name: doi
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reference metadata
          content:
            application/json:
              schema:
                type: object
        "404":
          description: DOI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Upstream lookup error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/references/lookup/isbn/{isbn}:
    get:
      tags: [References]
      summary: Lookup reference by ISBN
      description: Fetch reference metadata from OpenLibrary by ISBN.
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reference metadata
          content:
            application/json:
              schema:
                type: object
        "404":
          description: ISBN not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Upstream lookup error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/references/{id}:
    get:
      tags: [References]
      summary: Get a reference
      description: Retrieve a single reference by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reference object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "404":
          description: Reference not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [References]
      summary: Update a reference
      description: Update fields on an existing reference.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReferenceBody"
      responses:
        "200":
          description: Updated reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "403":
          description: No write access to reference library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Reference not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [References]
      summary: Delete a reference
      description: Remove a reference from the library.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "403":
          description: No delete access to reference library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Reference not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/audit/documents/{id}/log:
    get:
      tags: [Audit]
      summary: Get paginated audit log
      description: Retrieve the audit log for a document. Requires manage permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Array of audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/audit/verify/{documentId}:
    get:
      tags: [Audit]
      summary: Verify HMAC chain integrity
      description: Verify the HMAC chain integrity for a document's audit log. Requires manage permission.
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  entries:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/audit/proof/{documentId}:
    get:
      tags: [Audit]
      summary: Export full audit proof bundle
      description: Export a complete cryptographic audit proof for a document. Requires manage permission.
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full audit proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditProof"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/audit/proof/{documentId}/summary:
    get:
      tags: [Audit]
      summary: Export lightweight proof summary
      description: Export a lightweight summary of the audit proof for a document. Requires manage permission.
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Proof summary
          content:
            application/json:
              schema:
                type: object
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/audit/proof/verify:
    post:
      tags: [Audit]
      summary: Verify a proof bundle offline
      description: Stateless verification of an audit proof bundle. Requires auth but no document permissions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proof, hmacSecret]
              properties:
                proof:
                  $ref: "#/components/schemas/AuditProof"
                hmacSecret:
                  type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  entries:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
        "400":
          description: Missing proof or hmacSecret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ===== Workflow =====
...